backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
astroid==3.3.4
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cloudinary==1.44.0
contourpy==1.2.0
coverage==7.8.2
cryptography==45.0.2
cycler==0.12.1
dill==0.3.8
distlib==0.3.8
Django==5.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.0
Faker==30.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.13.1
fonttools==4.49.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.1
numpy==1.26.4
orjson==3.10.18
packaging==25.0
pandas==2.2.2
passlib==1.7.4
pbr==6.0.0
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.6.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pygame==2.5.2
Pygments==2.19.1
pylint==3.3.1
pylint-pytest==1.1.8
pymongo==4.13.0
pyparsing==3.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
redis==6.1.0
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.46.2
stevedore==5.1.0
tomli==2.0.1
tomlkit==0.13.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
watchfiles==1.0.5
websockets==15.0.1
